requirements.txt
cffi==1.14.0
colorcorrect==0.9
cryptography==2.8
Cython==0.29.14
decorator==4.3.0
-e git+https://github.com/NCI-GDC/gdc-client.git@1c69ed0c2bfa3c6b3784bca1ab6feaed7a81f6cb#egg=gdc_client
h5py==2.9.0
intervaltree==3.0.2
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
joblib==0.15.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
git+https://github.com/src-d/kmcuda.git#subdirectory=src
lxml==4.4.2
matplotlib==3.1.1
mygene==3.0.0
ndg-httpsclient==0.5.0
numba==0.45.1
numpy==1.17.0
openslide-python==1.1.1
pandas==0.23.4
pathlib==1.0.1
Pillow==6.1.0
progressbar2==3.43.1
pyasn1==0.4.3
pyOpenSSL==18.0.0
PyYAML==3.13
requests==2.22.0
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
setuptools==45.3.0
statsmodels==0.9.0
tables==3.5.2
tensorboard==1.14.0
tensorboardX==1.4
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.32.2
